Changed lookup table int to size_t (windows compilation issue), changed release flags to not kill on warnings (ubuntu clang issue with Catch2)

+++ src/sac_format.hpp
@@ -253,7 +253,7 @@ enum class name {
     data2
 };
 // Lookup maps
-const std::unordered_map<name, int> sac_map = {
+const std::unordered_map<name, size_t> sac_map = {
     // Floats
     {name::depmin,    0},
     {name::depmax,    1},
